debian/rules: work around lrelease failure when building under user-mode emulation.
In user-mode emulation, lrelease wrongly detects a cross-compiling
situation and fails to load the default mkspec for some reason.
Debugging this is incredibly frustrating, especially since strace and
gdb do not work correctly in user-mode emulation.

Fortunately, we can work around this issue by specifying QMAKESPEC
directly, ironically with values queried from qmake.

We will do that for Debian 12+ only for now. Note that this is not an
issue with full-machine emulation, since the inner system is
unsuspecting to the actual bare-metal host.