Makefile: pass actual values to variables in MAKEOVERRIDES and .MAKEFLAGS.
Not passing values leads to builds breaking with newer make versions,
since they are now passing MAKEOVERRIDES as-is, leading to build
failures since instead of exporting variables, the data passed is now
interpreted as a make target instead.

Passing values should (hopefully?) be compatible with older make
versions, though, and we would have needed to do this for .MAKEFLAGS the
whole time anyway.

Only .MAKEOVERRIDES on FreeBSD and NetBSD is special in that this
variable actually expects a variable list, without values.