Makefile: revert to the original behavior.
This unbreaks older GNU Make versions (up to 4.3), but leaves newer ones
(4.4 and up) broken.

This unfortunately includes Debian Unstable for now.

This is a regression in GNU Make and will need to be handled there,
since working around is nigh impossible (especially since values with
whitespace just can't be handled correctly).

It also means that OpenBSD's Make is essentially broken, but there's
nothing that can reasonably be done about this.